pyproject.toml
[project]
name = "python-bootcamp-project"
version = "0.1.0"
description = "YouTube Channel RAG Agent - Agent Engineering Bootcamp Project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "litellm>=1.72.9",
    "python-dotenv>=1.1.0",
    # YouTube RAG Dependencies
    "pytubefix>=6.16.4",
    "youtube-transcript-api>=0.6.2",
    "google-api-python-client>=2.141.0",
    "chromadb>=0.5.0",
    "langchain>=0.3.0",
    "langchain-chroma>=0.1.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.2.0",
    "streamlit>=1.38.0",
    "pandas>=2.2.0",
    "pyyaml>=6.0.0",
    "tqdm>=4.66.0",
    # MCP Integration Dependencies
    "mcp[cli]>=1.4.0",
    "httpx>=0.25.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "pytest>=8.4.1",
]
